Move Grid tests to separate file
